chore: trying new util convention
